+++ crates/swc_ecma_minifier/tests/fixture/issues/2257/full/output.js
@@ -7290,10 +7290,7 @@
                 return null === match ? {
                     value: void 0,
                     done: state.done = !0
-                } : state.global ? ("" === toString1(match[0]) && (R.lastIndex = advanceStringIndex(S, toLength(R.lastIndex), state.unicode)), {
-                    value: match,
-                    done: !1
-                }) : (state.done = !0, {
+                } : (state.global ? "" === toString1(match[0]) && (R.lastIndex = advanceStringIndex(S, toLength(R.lastIndex), state.unicode)) : state.done = !0, {
                     value: match,
                     done: !1
                 });
@@ -11055,7 +11052,6 @@
             }
             var kb = {
                 html: "http://www.w3.org/1999/xhtml",
-                mathml: "http://www.w3.org/1998/Math/MathML",
                 svg: "http://www.w3.org/2000/svg"
             };
             function lb(a) {
@@ -11422,7 +11418,7 @@
                 }
             }
             function tc(a, b, c, d, e, f) {
-                return null === a || a.nativeEvent !== f ? (a = rc(b, c, d, e, f), null !== b && null !== (b = Cb(b)) && fc(b), a) : (a.eventSystemFlags |= d, b = a.targetContainers, null !== e && -1 === b.indexOf(e) && b.push(e), a);
+                return null === a || a.nativeEvent !== f ? (a = rc(b, c, d, e, f), null !== b && null !== (b = Cb(b)) && fc(b)) : (a.eventSystemFlags |= d, b = a.targetContainers, null !== e && -1 === b.indexOf(e) && b.push(e)), a;
             }
             function xc(a) {
                 if (null !== a.blockedOn) return !1;
@@ -12889,16 +12885,16 @@
                     return a && null === b.alternate && (b.flags = 2), b;
                 }
                 function h(a, b, c, d) {
-                    return null === b || 6 !== b.tag ? ((b = Ug(c, a.mode, d)).return = a, b) : ((b = e(b, c)).return = a, b);
+                    return null === b || 6 !== b.tag ? (b = Ug(c, a.mode, d)).return = a : (b = e(b, c)).return = a, b;
                 }
                 function k(a, b, c, d) {
-                    return null !== b && b.elementType === c.type ? ((d = e(b, c.props)).ref = Qg(a, b, c), d.return = a, d) : ((d = Vg(c.type, c.key, c.props, null, a.mode, d)).ref = Qg(a, b, c), d.return = a, d);
+                    return null !== b && b.elementType === c.type ? (d = e(b, c.props)).ref = Qg(a, b, c) : (d = Vg(c.type, c.key, c.props, null, a.mode, d)).ref = Qg(a, b, c), d.return = a, d;
                 }
                 function l(a, b, c, d) {
-                    return null === b || 4 !== b.tag || b.stateNode.containerInfo !== c.containerInfo || b.stateNode.implementation !== c.implementation ? ((b = Wg(c, a.mode, d)).return = a, b) : ((b = e(b, c.children || [])).return = a, b);
+                    return null === b || 4 !== b.tag || b.stateNode.containerInfo !== c.containerInfo || b.stateNode.implementation !== c.implementation ? (b = Wg(c, a.mode, d)).return = a : (b = e(b, c.children || [])).return = a, b;
                 }
                 function n(a, b, c, d, f) {
-                    return null === b || 7 !== b.tag ? ((b = Xg(c, a.mode, d, f)).return = a, b) : ((b = e(b, c)).return = a, b);
+                    return null === b || 7 !== b.tag ? (b = Xg(c, a.mode, d, f)).return = a : (b = e(b, c)).return = a, b;
                 }
                 function A(a, b, c) {
                     if ("string" == typeof b || "number" == typeof b) return (b = Ug("" + b, a.mode, c)).return = a, b;
@@ -12986,7 +12982,7 @@
                             }
                             return g(a1);
                     }
-                    if ("string" == typeof f1 || "number" == typeof f1) return f1 = "" + f1, null !== d1 && 6 === d1.tag ? (c(a1, d1.sibling), (d1 = e(d1, f1)).return = a1, a1 = d1) : (c(a1, d1), (d1 = Ug(f1, a1.mode, h)).return = a1, a1 = d1), g(a1);
+                    if ("string" == typeof f1 || "number" == typeof f1) return f1 = "" + f1, null !== d1 && 6 === d1.tag ? (c(a1, d1.sibling), (d1 = e(d1, f1)).return = a1) : (c(a1, d1), (d1 = Ug(f1, a1.mode, h)).return = a1), g(a1 = d1);
                     if (Pg(f1)) return function(e, g, h, k) {
                         for(var l = null, t = null, u = g, z = g = 0, q = null; null !== u && z < h.length; z++){
                             u.index > z ? (q = u, u = null) : q = u.sibling;