+++ crates/swc_ecma_minifier/tests/fixture/projects/backbone/11/input.js
@@ -1,4 +1,4 @@
-const obj = {
+export const obj = {
     navigate: function (fragment, options) {
         if (!History.started) return false;
         if (!options || options === true) options = { trigger: !!options };