+++ crates/swc_ecma_minifier/tests/fixture/projects/backbone/5/output.js
@@ -1,4 +1,4 @@
-var Events = {
+export var Events = {
     on: function(name, callback, context) {
         return eventsApi(this, "on", name, [
             callback,